Extracted class suggestions logic into useClassSuggestions
